backend/requirements.txt
fastapi[all]
diffusers
transformers
accelerate
safetensors
omegaconf
pytorch_lightning
